+++ src/controller/controllerEmployee.ts
@@ -2,10 +2,9 @@ import { Response, Request } from 'express';
 import { Controller } from "../decorateur/controller";
 import { Delete, Get, Post, Put } from "../decorateur/route";
 import { executeQuery } from "../middelware/mysqlConnection";
-import { AddEmployee,  selectOneEmployee, updateOneEmployee, deleteOneEmployee, functionConnectionEmployee, creationTacheRequest } from "../middelware/requeteSqlEmployee";
+import { requeteSqlEmployee } from "../middelware/requeteSql/requeteSqlEmployee";
 import { schemaAddEmployee, schemaConnectionEmployee, schemaModifyEmployee } from "../schema/schemaEmployee";
-import { generateToken, verificationToken } from '../middelware/token/employeeToken';
-import { schemaCreationTache } from '../schema/schemaTacheEmployee';
+import { tokenClass } from '../middelware/token/employeeToken';
 
 
 @Controller()
@@ -21,7 +20,7 @@ class EmployeeController {
     async create(req: Request, res: Response) {
         try {
             await schemaAddEmployee(req, res);
-            AddEmployee(req.body.firstName, req.body.lastName, req.body.email, req.body.teams, req.body.password)
+            requeteSqlEmployee.AddEmployee(req.body.firstName, req.body.lastName, req.body.email, req.body.teams, req.body.password)
                 .then(() => { return res.status(201).json({message: "Employée ajouté avec success!"}) })
                 .catch(() => { return res.status(401).json({message: `Une erreur est survenue`}) })
         } catch (error) {
@@ -31,7 +30,7 @@ class EmployeeController {
 
     @Get("get-one/:id")
     async oneEmployee(req: Request, res: Response){
-        selectOneEmployee(req.params.id)
+        requeteSqlEmployee.selectOneEmployee(req.params.id)
             .then((result) => {
                 if (result.length === 0) {
                     return res.status(401).json({message: `L'employee n'existe pas!`})
@@ -46,7 +45,7 @@ class EmployeeController {
         try { 
             await schemaModifyEmployee(req, res)
             .then(() => {
-                updateOneEmployee(req.params.id, req.body.firstName, req.body.lastName, req.body.email, req.body.teams)
+                requeteSqlEmployee.updateOneEmployee(req.params.id, req.body.firstName, req.body.lastName, req.body.email, req.body.teams)
                     .then(() => { res.status(201).json({message: "L'employee a été modifié avec success!"}) })
                     .catch(() => { return res.status(401).json({message: `Une erreur est survenue`}) })
             })
@@ -58,7 +57,7 @@ class EmployeeController {
 
     @Delete('delete-employee/:id')
     async deleteOneEmployee(req: Request, res: Response){
-        deleteOneEmployee(req.params.id)
+        requeteSqlEmployee.deleteOneEmployee(req.params.id)
             .then(() => { res.status(201).json({message: "L'employee a été supprimé avec success!"}) })
             .catch(() => { return res.status(401).json({message: `Une erreur est survenue`}) })
     }
@@ -69,14 +68,14 @@ class EmployeeController {
         try {
             await schemaConnectionEmployee(req, res);
 
-            const result = await functionConnectionEmployee(req.body.email, req.body.password);
+            const result = await requeteSqlEmployee.ConnectionEmployee(req.body.email, req.body.password);
             const id = result[0].id
 
             if (!result) {
                 return res.status(401).json({ message: "L'un de vos identifiants n'est pas valide" });
             }
 
-            const token = await generateToken(id);
+            const token = await tokenClass.generateToken(id);
             return res.status(201).json({ message: "Vous êtes connecté", token: token });
         } catch (error) {
             console.error("Erreur lors de la connexion :", error);
@@ -84,21 +83,6 @@ class EmployeeController {
         }
     }
 
-    @Post('creation-tache')
-    async creationTache(req: Request, res: Response) {
-        const token = req.headers.authorization?.split(' ')[1]
-        try {
-            const tokenVerifier = await verificationToken(token);
-            await schemaCreationTache(req, res);
-            
-            creationTacheRequest(tokenVerifier?.id, req.body.assignerAquelEmployee, req.body.description)
-                .then(() => { return res.status(201).json({message: 'La tache a bien été ajouter'}) })
-                .catch((error) => { return res.status(500).json({message: 'Une erreur est survenue lors de la creation de la tache' + error}) })
-        } catch (error) {
-            return res.status(500).json({message: "Token invalid ou inactif!"})
-        }
-
-    }
 }
 
 export { EmployeeController }
